+++ god-service/src/main/java/com/god/service/zeus/GodUserService.java
@@ -36,7 +36,7 @@ public interface GodUserService {
 	 * @param data
 	 */
 	int add(GodUserBo godUserBo);
-
-	GodUser getGodUserById(Integer userId);
+	
+	GodUser getGodUserById(Long userId);
 
 }
